requirements.txt
sounddevice
cffi
numpy
anytree
appdirs
pyyaml
PySide6
soundfile
huggingface_hub
onnxruntime-gpu
